Extract SpotlightCard to dedicated shared widget file
- Move SpotlightCard to widgets/spotlight_card.dart for reuse
- Import from spotlight_card.dart in both results_step.dart and main.dart
- Clarifies that SpotlightCard is shared between ResultsStep and UsageStep
- Improves separation of concerns and reduces file coupling
